+++ src/usgs_adapter/__init__.py
@@ -1,0 +1,3 @@
+from usgs_adapter.client import SeismicEvent, USGSClient
+
+__all__ = ["SeismicEvent", "USGSClient"]